requirements.txt
asgiref==3.3.1
click==7.1.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.6
django-bootstrap3==14.2.0
django2-bootstrap3-datetimepicker==3.0.3
Flask==1.1.2
gunicorn==20.1.0
hupper==1.10.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PasteDeploy==2.1.1
Pillow==8.1.0
plaster==1.0
plaster-pastedeploy==0.7
pyramid==1.10.5
pytz==2021.1
reverse==0.1.0
sqlparse==0.4.1
static3==0.7.0
translationstring==1.4
venusian==3.0.0
WebOb==1.8.6
Werkzeug==1.0.1
zope.deprecation==4.4.0
zope.interface==5.2.0
psycopg2>=2.6.1
